Changes on the Warning section
The word "DISCLAIMER" is somehow meaningless if the section does not disclose the developer's exemption from liability if the end user damages their memory card, so it would be either reasonable to add the disclosure or remove the word from the section's title. Also, added extra info advising the reader that a method to restore the memory card's normal functionality is to overwrite the exploit using a certain game like Cool Boarders 4, which allows showing the card's content and selecting a block to write or overwrite the save file.